f5-traffic-simulator/requirements.txt
ansible>=6.0.0
pyyaml>=6.0
jinja2>=3.1.0
